requirements.txt
selenium==4.33.0
pytest==8.3.5
allure-pytest==2.14.2
webdriver-manager==4.0.2
